RFID-BasedMedicationTracker_Group7/requirements.txt
mfrc522-python
python-telegram-bot
requests
RPi.GPIO
schedule
spidev
